requirements.txt
requests
json
fake-useragent
logging
pandas
datetime
python-telegram-bot